requirements.txt
langchain>=0.3.0
langchain-text-splitters>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0
faiss-cpu>=1.8.0
pypdf>=5.0.0
sqlalchemy>=2.0.0
streamlit>=1.40.0
pandas>=2.2.0
faker>=28.0.0
python-dotenv>=1.0.1
mcp>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
reportlab>=4.0.0
psycopg2-binary>=2.9.10
psycopg2>=2.9.11
pytest>=9.0.3
